Added comprehensive SEO metadata across all public routes and noindex directives for authenticated pages.
Added SEO metadata, Open Graph tags, and a new robots.txt and sitemap to all public pages. Authenticated dashboard and admin pages are set to noindex so search engines don't index user data.

Closes #243
