+++ app/(dashboard)/dashboard-layout-client.tsx
@@ -1,0 +1,77 @@
+"use client";
+
+import { useEffect, useState } from "react";
+
+import { Sidebar } from "../../components/dashboard/sidebar";
+import { Topbar } from "../../components/dashboard/topbar";
+import { NetworkStatusBanner } from "@/components/shared/network-status-banner";
+import { cn } from "../../lib/utils";
+import { useAuthStore } from "../../hooks/use-auth-store";
+import { useRouter } from "next/navigation";
+import { useSidebarStore } from "../../hooks/use-sidebar-store";
+
+export default function DashboardLayoutClient({
+  children,
+}: {
+  children: React.ReactNode;
+}) {
+  const { isOpen, close } = useSidebarStore();
+  const [isSidebarCollapsed, setIsSidebarCollapsed] = useState(false);
+  const { isAuthenticated, accessToken } = useAuthStore();
+  const router = useRouter();
+
+  useEffect(() => {
+    if (!isAuthenticated || !accessToken) {
+      router.replace("/sign-in");
+    }
+  }, [isAuthenticated, accessToken, router]);
+
+  if (!isAuthenticated || !accessToken) {
+    return null;
+  }
+
+  return (
+    <div className="flex min-h-screen bg-background text-foreground transition-all duration-300">
+      {/* Sidebar - Desktop */}
+      <aside
+        className={cn(
+          "hidden md:block transition-all duration-300",
+          isSidebarCollapsed ? "w-20" : "w-64",
+        )}
+      >
+        <Sidebar
+          isCollapsed={isSidebarCollapsed}
+          onToggleCollapse={() => setIsSidebarCollapsed(!isSidebarCollapsed)}
+        />
+      </aside>
+
+      {/* Sidebar - Mobile Drawer Overlay */}
+      {isOpen && (
+        <div
+          className="fixed inset-0 z-40 bg-black/50 md:hidden animate-in fade-in duration-300"
+          onClick={close}
+        />
+      )}
+
+      {/* Sidebar - Mobile Drawer */}
+      <aside
+        className={cn(
+          "fixed inset-y-0 left-0 z-50 w-70 transform transition-transform duration-300 ease-in-out md:hidden bg-white dark:bg-black",
+          isOpen ? "translate-x-0" : "-translate-x-full",
+        )}
+      >
+        <Sidebar />
+      </aside>
+
+      <div className="flex flex-1 flex-col overflow-hidden">
+        <NetworkStatusBanner />
+        <div className="p-4 md:px-8">
+          <Topbar />
+        </div>
+        <main className="flex-1 overflow-y-auto px-4 md:px-8 pb-4">
+          {children}
+        </main>
+      </div>
+    </div>
+  );
+}
