+++ app/layout.tsx
@@ -24,8 +24,21 @@ const manrope = Manrope({
 });
 
 export const metadata: Metadata = {
-  title: "NexaFx",
-  description: "Multi-currency finance on Stellar",
+  title: {
+    default: "NexaFx — Multi-Currency Finance on Stellar",
+    template: "%s | NexaFx",
+  },
+  description:
+    "Convert, deposit, and transfer currencies instantly on the Stellar blockchain.",
+  keywords: [
+    "currency exchange",
+    "Stellar blockchain",
+    "cross-border payments",
+    "NGN to USD",
+    "crypto finance",
+  ],
+  authors: [{ name: "Nexacore" }],
+  creator: "Nexacore",
   manifest: "/manifest.json",
   appleWebApp: {
     capable: true,
@@ -37,6 +50,26 @@ export const metadata: Metadata = {
       { url: "/icons/icon-152x152.png", sizes: "152x152", type: "image/png" },
     ],
   },
+  openGraph: {
+    type: "website",
+    locale: "en_US",
+    url: "https://nexafx.io",
+    siteName: "NexaFx",
+    title: "NexaFx — Multi-Currency Finance on Stellar",
+    description:
+      "Convert, deposit, and transfer currencies instantly on the Stellar blockchain.",
+    images: [
+      { url: "/og-image.png", width: 1200, height: 630, alt: "NexaFx" },
+    ],
+  },
+  twitter: {
+    card: "summary_large_image",
+    title: "NexaFx — Multi-Currency Finance on Stellar",
+    description:
+      "Convert, deposit, and transfer currencies instantly on the Stellar blockchain.",
+    images: ["/og-image.png"],
+  },
+  robots: { index: true, follow: true },
 };
 
 export const viewport: Viewport = {
