+++ package.json
@@ -18,6 +18,7 @@
     "@radix-ui/react-tabs": "^1.1.15",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "eslint-plugin-storybook": "^10.4.6",
     "lucide-react": "^1.21.0",
     "next": "16.2.9",
     "next-pwa": "^5.6.0",
